build: Use universal workspace-relative path to TCF file
This commit replaces absolute path to a TCF file with an Eclipse
variable expression `${workspace_loc:...}`, which allows this path to
work on any host out of the box. To work properly this commit requires
change in GNU IDE, otherwise path will not be expanded:

    2b5ff58 elf32: Allow TCF path to be Eclipse variable expression

However current content of .cproject is not any better, since that path is
specific to a certain host and will not work anywhere else.